+++ tests/test_libudev.py
@@ -41,6 +41,8 @@ WRAPPER_BLACKLIST_PATTERNS = [
     'udev_list_entry_get_by_name',
     # superfluous because of ".encode('string-escape')"
     'udev_util_encode_string',
+    # deprecated and removed in recent udev versions
+    'udev_monitor_new_from_socket',
     # all queue functions (queue interface is not wrapped)
     re.compile('^udev_queue_.*'),